Refactor: move constraint code into a separate file
